fix: remove alias arg instead of opt
