+++ tests/integration/test_pending_cli.py
@@ -1,0 +1,3 @@
+def test_help(runner, cli):
+    result = runner.invoke(cli, ["--help"], catch_exceptions=False)
+    assert result.exit_code == 0, result.output
